Fix product removal undo

+++ src/store/modules/removebranch.js
@@ -319,38 +319,39 @@ const actions = {
 						id: nowSelectedNode._id, onSuccessCallback: () => {
 							const removedNode = payload.node
 							// remove the node and its children from the tree view
-							window.slVueTree.removeNodes([removedNode])
-							// remove the children; on restore the children are recovered from the database
-							removedNode.children = []
-							if (removedNode.level === LEVEL.PRODUCT) {
-								// remove the product from the users product roles, subscriptions and product selection array and update the user's profile
-								dispatch('removeFromMyProducts', { productId: removedNode._id })
-							}
+							if (window.slVueTree.removeNodes([removedNode])) {
+								// remove the children; on restore the children are recovered from the database
+								removedNode.children = []
+								if (removedNode.level === LEVEL.PRODUCT) {
+									// remove the product from the users product roles, subscriptions and product selection array and update the user's profile
+									dispatch('removeFromMyProducts', { productId: removedNode._id })
+								}
 
-							if (payload.createUndo) {
-								const removedDescendantsCount = docsRemovedIds.length - 1
-								// create an entry for undoing the remove in a last-in first-out sequence
-								const entry = {
-									type: 'undoRemove',
-									delmark: payload.delmark,
-									isProductRemoved: removedNode.level === LEVEL.PRODUCT,
-									itemsRemovedFromReqArea,
-									removedDescendantsCount,
-									removedExtConditions: removed.removedExtConditions,
-									removedExtDependencies: removed.removedExtDependencies,
-									removedIntConditions: removed.removedIntConditions,
-									removedIntDependencies: removed.removedIntDependencies,
-									removedNode,
-									sprintIds: removedSprintIds
+								if (payload.createUndo) {
+									const removedDescendantsCount = docsRemovedIds.length - 1
+									// create an entry for undoing the remove in a last-in first-out sequence
+									const entry = {
+										type: 'undoRemove',
+										delmark: payload.delmark,
+										isProductRemoved: removedNode.level === LEVEL.PRODUCT,
+										itemsRemovedFromReqArea,
+										removedDescendantsCount,
+										removedExtConditions: removed.removedExtConditions,
+										removedExtDependencies: removed.removedExtDependencies,
+										removedIntConditions: removed.removedIntConditions,
+										removedIntDependencies: removed.removedIntDependencies,
+										removedNode,
+										sprintIds: removedSprintIds
+									}
+									if (entry.isProductRemoved) {
+										entry.removedProductRoles = rootGetters.getMyProductsRoles[removedNode._id]
+									}
+									rootState.changeHistory.unshift(entry)
+									commit('showLastEvent', { txt: `The ${getLevelText(rootState.configData, removedNode.level)} and ${removedDescendantsCount} descendants are removed`, severity: SEV.INFO })
+								} else {
+									commit('showLastEvent', { txt: 'Item creation is undone', severity: SEV.INFO })
 								}
-								if (entry.isProductRemoved) {
-									entry.removedProductRoles = rootGetters.getMyProductsRoles[removedNode._id]
-								}
-								rootState.changeHistory.unshift(entry)
-								commit('showLastEvent', { txt: `The ${getLevelText(rootState.configData, removedNode.level)} and ${removedDescendantsCount} descendants are removed`, severity: SEV.INFO })
-							} else {
-								commit('showLastEvent', { txt: 'Item creation is undone', severity: SEV.INFO })
-							}
+							} else commit('showLastEvent', { txt: `Cannot remove remove node with title ${removedNode.title}`, severity: SEV.ERROR })
 						}
 					})
 				}
